Move regression-related end-to-end Datalab tests to test_regression module
